+++ src/Spec.cpp
@@ -62,6 +62,10 @@ vector<SpecMatch> Spec::match(SEXP obj) const {
           continue;
         SEXP nm = has_names ? STRING_ELT(obj_names, ix) : R_NilValue;
         out.emplace_back(ix, name, nm, VECTOR_ELT(obj, ix));
+      } else {
+        if (defsexp != R_NilValue) {
+          out.emplace_back(ix, name, R_NilValue, defsexp);
+        }
       }
     }
   } else if (exclude_ixes.size() > 0) {
@@ -78,17 +82,30 @@ vector<SpecMatch> Spec::match(SEXP obj) const {
     if (exclude_names.size() > 0)
       out.reserve(N);
 
+    vector<bool> processed(include_names.size(), false);
+
     for (int i = 0; i < N; i++) {
       SEXP nm = R_NilValue;
       nm = STRING_ELT(obj_names, i);
-      if (is_char_in_strvec(nm, exclude_names)) {
-        continue;
-      } else if (include_names.size() > 0) {
-        if (!is_char_in_strvec(nm, include_names)) {
-          continue;
+      if (ix_char_in_strvec(nm, exclude_names) < 0) {
+        if (include_names.size() > 0) {
+          R_xlen_t ix = ix_char_in_strvec(nm, include_names);
+          if (ix >= 0) {
+            out.emplace_back(i, name, nm, VECTOR_ELT(obj, i));
+            processed[ix] = true;
+          }
+        } else {
+          out.emplace_back(i, name, nm, VECTOR_ELT(obj, i));
         }
       }
-      out.emplace_back(i, name, nm, VECTOR_ELT(obj, i));
+    }
+
+    if (defsexp != R_NilValue) {
+      for (size_t i = 0; i < processed.size(); i++) {
+        if (!processed[i]) {
+          out.emplace_back(-1, name, include_names[i], defsexp);
+        }
+      }
     }
   }
 
@@ -136,7 +153,8 @@ Spec sexp2spec(SEXP lspec) {
   bool done_as = false,
     done_children = false, done_groups = false,
     done_stack = false, done_process = false,
-    done_include = false, done_exclude = false;
+    done_include = false, done_exclude = false,
+    done_default = false;
   SEXP children = R_NilValue, groups = R_NilValue;
 
   Spec spec;
@@ -160,6 +178,9 @@ Spec sexp2spec(SEXP lspec) {
       } else if (!done_process && !strcmp(nm, "process")) {
         spec.process = sexp2process(obj);
         done_process = true;
+      } else if (!done_default && !strcmp(nm, "default")) {
+        spec.defsexp = obj;
+        done_default = true;
       } else if (!done_children && !strcmp(nm, "children")) {
         if (TYPEOF(obj) != VECSXP)
           Rf_error("spec's 'children' field must be a list");