+++ NEWS.md
@@ -1,6 +1,7 @@
 Version 0.0.2.9000 [Dev]
 ========================
 
+- New `default` argument in `s()` specification.
 - Data frame columns are now stacked by default (`stack_atomic = NULL` parameter)
 
 Version 0.0.2